--- a/script/unit-tests.ts
+++ b/script/unit-tests.ts
@@ -13,7 +13,7 @@ const environmentVariables = {
   TEST_ENV: '1',
   // ensuring Electron doesn't attach to the current console session (Windows only)
   ELECTRON_NO_ATTACH_CONSOLE: '1',
-  // speed up ts-node usage by using the new transpile-only module
+  // speed up ts-node usage by using the new transpile-only mode
   TS_NODE_TRANSPILE_ONLY: 'true',
 }