Currently our unit tests seem to spend a significant time transpiling on some platforms, e.g. [AppVeyor](https://ci.appveyor.com/project/github-windows/desktop/build/10153):

```
  ...

  331 passing (39s)
  8 pending
Done in 413.07s.
```

And [VSTS (Windows agent)](https://github.visualstudio.com/Desktop/_build/results?buildId=4566&view=logs):

```
[92m [0m[32m 331 passing[0m[90m (49s)[0m
 [36m [0m[36m 8 pending[0m
... 
Done in 449.09s.
```

We've upgraded `ts-node` recently, so perhaps there's some flags we need to set in here. An obvious one I was just looking at was `--transpileOnly`:

> `-T`, `--transpileOnly` Use TypeScript's faster `transpileModule` (`TS_NODE_TRANSPILE_ONLY`, default: `false`)

I'll update this with more research and digging.