@shiftkey 2018-08-13
Yep, this is a great win. Appveyor is now:

```
  331 passing (42s)
  8 pending
Done in 62.01s.
```

And VSTS on Windows is now:

```
[92m [0m[32m 331 passing[0m[90m (51s)[0m
[36m [0m[36m 8 pending[0m
...
Done in 81.71s.
```