@shiftkey 2018-08-15
Friendly bump on this - I talked with @outofambit yesterday and we're aiming to ship to production Thursday morning EST so we can both be on deck for monitoring throughout the day:

> I'd love for this to be merged before 1.3.3 ships as this cuts several minutes from each CI build, and does not touch app code, but feel free to push back if you have reservations.